fix: move invalid_union handler before message-based guards in formatZodIssue
In Zod v4, invalid_union issues have a default message of "Invalid input",
which caused two message-based guards to intercept union errors before
the invalid_union branch could run, producing degraded output.

Move the issue.code === "invalid_union" check to the top of formatZodIssue
so it always takes priority over string-matching fallbacks.

Also consolidate zod tests: merge src/lib/errors/zod.test.ts into
src/lib/errors/__tests__/zod.test.ts (adding the two invalid_union
test cases) and remove the top-level test file.
