+++ src/lib/errors/zod.ts
@@ -22,7 +22,21 @@ export class ZodValidationError extends BaseError {
 }
 
 function formatZodIssue(issue: z.ZodIssue): string {
-  // For Zod v4, let's use the default message but improve some specific cases
+  // Check by issue.code first (before any message-based string matching)
+  if (issue.code === "invalid_union") {
+    const branches = issue.errors
+      .map((branchErrors, i) => {
+        const firstIssue = branchErrors[0];
+        if (!firstIssue) return `option ${i + 1}`;
+        const branchPath =
+          firstIssue.path.length > 0 ? ` (missing: ${firstIssue.path.join(".")})` : "";
+        return `option ${i + 1}${branchPath}: ${firstIssue.message}`;
+      })
+      .join("; ");
+    return `Value doesn't match any expected format. Expected one of: [${branches}]`;
+  }
+
+  // Existing message-based checks follow...
   const message = issue.message;
 
   // Customize common error patterns for better UX
@@ -83,19 +97,6 @@ function formatZodIssue(issue: z.ZodIssue): string {
       return `Expected ${typeMatch[1]}, but received ${typeMatch[2]}`;
     }
   }
-  if (issue.code === "invalid_union") {
-    const branches = issue.errors
-      .map((branchIssues, i) => {
-        const firstIssue = branchIssues[0];
-        if (!firstIssue) return `option ${i + 1}`;
-        const branchPath =
-          firstIssue.path.length > 0 ? ` (missing: ${firstIssue.path.join(".")})` : "";
-        return `option ${i + 1}${branchPath}: ${firstIssue.message}`;
-      })
-      .join("; ");
-    return `Value doesn't match any expected format. Expected one of: [${branches}]`;
-  }
-
   if (message === "Invalid input") {
     return "Value doesn't match any of the expected types";
   }
